+++ src/alphadb/utils/query/table/altertable.py
@@ -16,8 +16,9 @@
 from alphadb.utils.types import Database
 from alphadb.utils.common import convert_version_number
 from alphadb.utils.concatenate.primary_key import get_primary_key
-from alphadb.utils.concatenate.column import get_column_renames
+from alphadb.utils.concatenate.column import get_column_renames, concatenate_column
 from alphadb.utils.query.column.addcolumn import addcolumn
+from alphadb.utils.query.column.definecolumn import definecolumn, prepare_definecolumn_data
 
 def altertable(version_source: dict, table_name: str, version: str, engine: Database):
     query = f" ALTER TABLE {table_name}"
@@ -79,14 +80,14 @@ def altertable(version_source: dict, table_name: str, version: str, engine: Data
             else:
                 this_column = table_data["modifycolumn"]
 
-            column_data = prepare_create_column_data(table_name, column, this_column, version, engine)
+            column_data = prepare_definecolumn_data(table_name, column, this_column, version, engine)
 
             #### If column data is None, its some attribute that should be handled later (foreign_key, primary_key, etc...)
             if column_data == None: continue
             
             if engine == "postgres": query += " ALTER COLUMN"
             else: query += " MODIFY COLUMN"
-            query += create_table_column(column_name=column, column_type=this_column[column]["type"], submethod="modifycolumn", length=column_data["length"], null=column_data["null"], unique=column_data["unique"], default=column_data["default"], auto_increment=column_data["auto_increment"], engine=engine)
+            query += definecolumn(column_name=column, column_type=this_column[column]["type"], submethod="modifycolumn", length=column_data["length"], null=column_data["null"], unique=column_data["unique"], default=column_data["default"], auto_increment=column_data["auto_increment"], engine=engine)
             query += ","
 
     #### Rename column
